Code review - extract
